Add a version support table to the FAQ
Fixes #15582

List the minor Teleport versions we currently support, plus their
release dates and minimum `tsh` versions. This restores the table that
we removed in PR #12641. Unlike that table, this change does not
list versions that are EOL, as this is now quite a lengthy list.
Instead, this change makes it explicit that the table lists currently
supported versions.