Add omitempty for GitHub teams_to_roles
Similar to #15993, in Teleport 10 we accept both teams_to_logins
and teams_to_roles. Without omitempty, Teleport will always encode
a null field on the connector YAML for whichever field is not
specified.

This is not an issue for v11+, as we will only support teams_to_roles
in Teleport 11.