Prevent ssh.Session SendRequest from wrapping payload twice
Payloads were inadvertently being wrapped twice by sessions making
the ssh server unable to parse the request payload. This is largely
only an issue for window-change requests, as they are the only request
being sent directly via a session that contains a payload.

As a means to migrate away from sending window-change requests directly
and to using the builtin ssh.Session WindowChange function this updates
the WinChangeReqParams to contain the values that will be sent in
accordance with RFC 4254 once we leverage the WindowChange function
instead of sending the request ourselves.