Fix infinite session heartbeat failures (#16065)
The session heartbeat loop wasn't checking if the context was
done, which resulted in an infinite loop of UpdateSession calls
that would never succeed and constantly trip the circuit breaker.

The context will only be cancelled if the session is terminated, so
we can simply return when either the context is done or the stop
channel was closed, whichever comes first.

Fixes #15695