[v9] Add OS compatibility checker (#16141)
* Add binary compatibility checking script (#14539)

Co-authored-by: Marek Smoliński <marek@goteleport.com>

* Adds CI hooks for GLibc compatibility check (#15547)

Adds the GCB build yaml for controlling the build, and updates the test script
to work in both the GCB environment and on a local dev machine.

Also changes the centos buildbox to leave the default user as root. When
GCB mounts the workspace into the container, the source code is owned
by root, and there is no way to change this. This means that the build will
fail when the non-root user specified in the build image attempt to write files
into the workspace. Setting the root user fixes this.

See-Also: #15186

* Serialise OS Compat check build (#16032)

The OS compat check is failing intermittently, and the parallel build is making it hard to diagnose the underlying cause. This patch reverts to a serialized build, making failure points more obvious.

Co-authored-by: Marek Smoliński <marek@goteleport.com>
Co-authored-by: Trent Clarke <trent@goteleport.com>
	git update-ref -d CHERRY_PICK_HEAD