Move old port allocation back into `integrations`
The old port allocation system is deprecated and should not be used
either by new codeor code in packages other than integrations. If this
code is use by more than one package, the potential for a port clas
rises, as each package will be drawing from the same range of port
numbers.

To make it harder (ideally impossible) for the old allocation system
to be used inappropriately, this patch moves it back into the
`integrations` packkages and marks it private.