operator: Fix flaky drift tests (#16269)
Fixes #15815

The drift tests were not suspending the operator reconciliation loop when causing the drift. If the operator reconciles before during the drift setup, we are not testing what we want (the operator creates the resource back in Teleport).

This PR adds the methods `startKubernetesOperator` and `stopKubernetesOperator` on the `testSetup` struct to fix drift tests.