vagrant: Remove reference to non-existing VM in deprecated S3 bucket
The `s3.gravitational.io` S3 bucket is going away, and this particular
VM has already been long-deleted from it. Replace it with the same
Debian VM image used for the `virtualbox`.

This hasn't been tested, so not sure it works as an equal replacement,
but seemed like the best next thing. Alternatively, could remove this
provider entirely.