Add Default Allow Rules for new resources (#16237)
* Add Default Allow Rules for new resources

When adding a new resource, we usually also set some default access
rules on the preset Roles (editor, access, auditor).

However, for existing clusters those new rules are not applied when we
start the cluster with the new teleport version.

This then requires manual intervention for setting up access to those
resources.

This PR adds a new mechanism to remove the need for manual intervention.

For each of the preset rules, we'll have default rules.
When the server starts, it tries to create the preset roles but if they
exist, instead of failing we will:
- check if the resource exists either in allow or deny rules
  - if so, then we don't do anything else
- add the default rule

This assumes it's ok to re-create the default rules when the admin
removes those resources from the allow/deny rule set.

* add log when adding default allow rules

Co-authored-by: Roman Tkachenko <roman@goteleport.com>