docs: move S3 IAM policy into an include
Prior to this change, the only place you could find a complete IAM
policy for using S3 to store session recordings was in the helm
deployment section.

Move the IAM policy into an include so that we can also include it
in the S3 section of the backends reference page.