Add a timeout for device cancels (#16575)
U2F devices, like Yubikey 4, don't seem to respect cancels, locking the program
if a different security key is chosen.

There isn't a clean way to cancel when Cancel() doesn't work, so I've added a
timeout to let the program complete. It's OK for the timeout to be short, it's
not tied to user interaction and by the time we get to it all cancels were
already issued.