Add a version support table to the FAQ (#15924) (#16630)
* Add a version support table to the FAQ

Fixes #15582

List the minor Teleport versions we currently support, plus their
release dates and minimum `tsh` versions. This restores the table that
we removed in PR #12641. Unlike that table, this change does not
list versions that are EOL, as this is now quite a lengthy list.
Instead, this change makes it explicit that the table lists currently
supported versions.

* Respond to PR feedback

Add a note about cross-component version compatibility to the FAQ by
creating a partial from a similar compatibility note in our Upgrading
guide.

Co-authored-by: Paul Gottschling <paul.gottschling@goteleport.com>