Use proper context in trackSession (#16687)
trackSession was always using `s.serverCtx` instead of the passed
in context, which prevented having spans from being added to the
correct trace. Calls to NewSessionTracker now use the proper request
scoped context. The backgroun UpdateExpirationLoop still uses
`s.serverCtx`, but the span created for it is properly linked to
the request context.