Stop checking the cert's province field for kube groups
In Teleport 9 and earlier, we would encode the kubernetes groups in
both a dedicated ASN.1 extension as well as the province field.

Starting in Teleport 10, we use the ASN.1 extension only and do not
set the province field. Teleport 10 still supported reading from the
province field in case it was using certs that were issued by an
older version of Teleport.

For Teleport 11, we stop checking the province field entirely.