Move SAML/OIDC web handlers to enterprise repository
Remove the SAML and OIDC web handlers from this repository as they have
been added to the enterprise repository. The SAML and OIDC connectors
are not available in the OSS edition and connectors of these types could
not be created.

This change should not have any effect on existing OSS deployments.

NOTE: This commit cannot be merged until the corresponding commit in the
enterprise repository is merged and the commit hash of the `e` submodule
updated to correspond to that of the merged commit in the enterprise
repository.