Properly get client address for app session audit.
The create app session audit log now properly gets the client address by
passing the client address in the gRPC messages to Teleport. The peer
retrieval that was being used previously was getting the address between
the proxy and the auth server, which was not displaying the same value as before.