Directory sharing specific functions are blocked by the go-side of the rdp client if directory sharing is disabled, however because of how that client is written, its relatively difficult to test for in go. (See https://github.com/gravitational/teleport/blob/bc81db46558d26011abe149d43949aa783dad2b4/lib/srv/desktop/rdp/rdpclient/client.go#L406-L416).
This commit elects to ignore that difficulty by adding a check on the
rust-side of the client, and testing that it works properly. In order to
do so, I needed to refactor the client slightly in order to get
`mcs::Client<S>` out of `rdpdr.rs`, because it was difficult to mock.