Fix test-go-root make target and tbot init tests (#16870)
Fix the `test-go-root` target in the `Makefile`, adding a missing
backslash at the end of the first line of the recipe.

Use `User.Username` instead of `User.Name` in the tbot
`TestInitMaybeACLs` test as that value ultimately is passed to
`user.Lookup`, which requires a username. I assume tests on CI were
running where Username == Name.