Store `host_uuid` into Kubernetes Secret
When recovering after a restart, `host_uuid` changes and Auth Server authentication fails.
This happens because `host_uuid` is not stored in the Kubernetes Secret but it's stored in at
the certificate Common Name.
This PR forces the storage and of the `host_uuid` into Kubernetes Secrets for later reuse.

Fixes #17474