Allow users to specify default values for Kubernetes impersonated users and groups.
When multiple Kubernetes users are defined for the same Teleport user he has to choose one using the `--as` `kubectl` flag.
If he does not specify it, Teleport will reject the request.

This PR adds an option to specify the `--as` and `--as-groups` directly into the kubeconfig while the user logs in into the cluster.
It also changes the `tsh kube login kubeCluster` behaviour. Previously, it dumped every cluster the user had access to in Teleport and selected the context for `kubeCluster`. This is no longer the behavior and `tsh kube login kubeCLuster` will just update or append the entry for `kubeCluster` in the `kubeconfig`.

Fixes #11402