Allow connections to node when auth is offline
The connection flow for `tsh ssh user@foo` was to first check if
per-session mfa was required, issue certificates if necesarry, and
then attempt to connect to the target. In the event that the first
step failed due to intermittent or no connectivity to auth the
connection attempt would end there. However, the check to auth is
not strictly required though. If the node is still reachable it
will determine whether the users existing cerfiticate permits
access or whether the user is required to perform per-session MFA.

By making `tsh ssh user@foo` always attempt to connect to the
target node even if it can't be certain whether or not per-session
MFA is required users will still be able to establish a session
in the event auth is unreachable. For users without per-session
MFA this means they will always have access to a node if it is
reachable. For users with per-session MFA this means in the worst
case they attempt to connect to the node and are returned an
AccessDenied error and will have to try again when the auth
server is reachable to perform the MFA ceremony.

To achieve this, `tsh` now utilizes `TeleportClient.ConnectToNode`,
which manages the first part of the initial flow and then calls
either `ProxyClient.ConnectToNode` or ProxyClient.PortForwardToNode`.
Neither of these perform any of the MFA ceremony anymore and must now
explicitly be passed the `ssh.AuthMethods` to be used for the connection.
If `TeleportClient.ConnectToNode` is able to talk to auth the
`ssh.AuthMethods` will be the certificates received from the MFA
ceremony, otherwise they will contain the already established
credentials.

Included is also a minor refactoring to how things are passed into
`runCommandOnNodes` as the siteName provided was always the site
from the `TeleportClient` which we already have access to since
its a receiver method.