[v11] [Connect] Add Server Features to GetCluster #17497 (#17642)
This will set the foundation for feature detection use in the Connect application. Previously, we could get all the information we needed just from the cluster profile (stored on disk) but now, with the demand to understand which features may be available on the current Auth server, we must ping. If there is a better way to handle this, I'm all ears!

Currently, the only relevant information needed for Connect is AdvancedAccessWorkflows (access requests), but we can add keys to the Connect API as needed.