Move SAML/OIDC web handlers from OSS repository (#562)
Move the SAML and OIDC web handlers from the Teleport OSS repository, as
these features are only available in the enterprise edition of Teleport.

This is the first step in moving the SAML and OIDC login logic from the
OSS repo. Follow-up commits will move the auth flows from the `lib/auth`
package.

This code added in this commit is identical to the code removed from the
OSS repository, with a few minor changes:

* Some symbols have had `web.` prefixed as they were in the same
  package as the code before, but now are not.
* The type on which the methods are defined has changed
* Access to the `Logger` and `ProxyClient` are slightly different.

However, these changes are mechanical, and no alteration to the logic
was made.

The test case `TestSAML` has also been moved across. This required
duplicating the `webSuite` used in the original package, stripping out
all the parts that are not needed for testing SAML.

Source-commit: gravitational/teleport.e@7789b03f7233988106a76af4e1333e7ffe87572b
Source-pull-requests: gravitational/teleport.e#562