Move SAML/OIDC web handlers to enterprise repository (#17459)
* Move SAML/OIDC web handlers to enterprise repository

Remove the SAML and OIDC web handlers from this repository as they have
been added to the enterprise repository. The SAML and OIDC connectors
are not available in the OSS edition and connectors of these types could
not be created.

This change does not have any effect on existing OSS deployments.