[v11] Connect: Accommodate for making gRPC client creds from tshd key pair (#17841)
For tshd-initiated communication, the tshd process will need to create a
client that will connect to a gRPC server operated by the renderer
process of the Electron app.

On Windows, we use gRPC over TCP with mTLS. Each process creates its own
keypair and saves the public key to a predetermined location.

The previous code assumes that tshd is only going to need server
credentials. This commit makes it possible to create client credentials
from the same key pair.