[v11] Connect: Accommodate for making gRPC server creds in shared process (#1220) (#1302)
* Add comment to resolveNetworkAddress

* Log errors during boot of React app

* Connect: Accommodate for making gRPC server creds in shared process

* Adjust timeout for resolveNetworkAddress

Some of the timeouts in preload.ts, the shared process and tshd are set
to 10s. If the timeout to resolve the network address is lower or equal
to that, those timeout won't have the chance kick in.

Hence setting the timeout to a slightly larger value.

* makeCert: Add subjectAltName to certs

Without this, the cert generated by makeCert was failing when loaded
in Go as the gRPC server cert. The error message was:

    x509: certificate relies on legacy Common Name field, use SANs instead

See:

* https://stackoverflow.com/a/31143907/742872
* Search for "subjectAltName" https://github.com/digitalbazaar/forge/blob/2bb97afb5058285ef09bcf1d04d6bd6b87cffd58/README.md

* Fix lint issues

* Reword comment

Co-authored-by: Michael <michael.myers@goteleport.com>

* Remove unused imports

Co-authored-by: Michael <michael.myers@goteleport.com>

Co-authored-by: Michael <michael.myers@goteleport.com>