Refactor (*TermManager) Read
Removes spawning a goroutine per `Read` and eliminates using a `sync.Cond`
to indicate whether data flow is on/off. Doing so results in a
goroutine **per character** written that lasts for the entire duration
of the session. A large number of sessions, or very long lived sessions
caused the Proxy to consume large amounts of memory.

The mechanism was meant to prevent reading if data flow was not enabled
for a session due to certain constraints of moderated sessions. By replacing
the `sync.Cond` with a `chan bool` the same effect can be achieved
without needing a goroutine.

Fixes #17666