helm: Handle setting teleportVersionOverride to <11 when using v11 teleport-kube-agent chart
The instructions for adding a teleport-kube-agent in Teleport's web UI always use the latest version of the Helm
chart from the repo. This works fine when installing v11 agents, but causes errors if you set `teleportVersionOverride`
to anything <11 because config version `v3` is not supported. This will cause issues for Teleport Cloud customers trying
to add new agents following the web UI's instructions, as Cloud is not on v11 yet and won't be for a few weeks.

The behaviour of `semverCompare` is pretty awful and Teleport version convention dictates that we always use `-dev`, `-rc`
or other similar suffixes for versioning. As such, we split the string and take the first part to use for a semver comparison.