helm: Handle setting teleportVersionOverride to <11 when using v11 teleport-kube-agent chart (#18008)
* helm: Handle setting teleportVersionOverride to <11 when using v11 teleport-kube-agent chart

The instructions for adding a teleport-kube-agent in Teleport's web UI always use the latest version of the Helm
chart from the repo. This works fine when installing v11 agents, but causes errors if you set `teleportVersionOverride`
to anything <11 because config version `v3` is not supported. This will cause issues for Teleport Cloud customers trying
to add new agents following the web UI's instructions, as Cloud is not on v11 yet and won't be for a few weeks.