Fixes TestListKube when running in Github Actions
When running tests in Github Actions, Teleport automatically imports the instance labels as service labels.
This results in the TestListKube test to fail since it receives unexpected labels.
This PR replaces the default IMDSClient with a fake one to avoid introducing labels imported from the CI env where the test is running.
It also removes an unrelated error message - self subject access review failed - that is printed but had no impact on the test result.