Fix caching of client credentials when users re-login to incorporate new attributes
When users re-login after a failed attempt to access a Kubernetes cluster, Teleport may continue to use the old credentials for cluster access. This behavior results in successive failures until the credential cache expires (~1h).

This PR incorporates the changes made by @r0mant to solve the caching problem by incorporating certificate expiration into the cache key. So every time the user re-login, the key will be different because the expiration date of the certificate will be different and thus will not use the cached entry.

Fixes #18070