Restore the Ansible Server Access guide
Closes #14692

PR #10987 repurposed the Ansible Server Access guide to be a Machine ID
guide. Since users still found use in the original Server Access guide,
this change restores that guide.

Note that this change does not attempt to edit the Ansible Server Access
guide, and simply restores it to the state it was in before #10987. The
Machine ID Ansible guide is left untouched.