[v11] Allow users to specify default values for Kubernetes impersonated user, groups and namespace (#18185)
This PR adds an option to specify the `--as`, `--as-groups`, and `--kube-namespace|-n` directly into the kubeconfig while the user logs in to the cluster. 

It also changes the `tsh kube login kubeCluster` behavior. Previously, it dumped every cluster the user had access to in Teleport and selected the context for `kubeCluster`. This is no longer the behavior and `tsh kube login kubeCLuster` will just update or append the entry for `kubeCluster` in the `kubeconfig`.

Fixes #11402
Fixes #17640