auth: Export ssoDiagContext for external use
Export the struct `ssoDiagContext` (now `SSODiagContext`) and its fields
and methods and decouple it from `auth.Server`. This is a helper type
for recording SSO test diagnostics to simplify passing SSO context
between the various SSO methods. It contains very little functionality
in its own right and does not expose much if any of the auth.Server
implementation by exporting it.

The decoupling is achieved by defining a small interface (one function)
for what is required by `SSODiagContext` to record the diagnostic info
in the backend. The `auth.Server` implements this interface already, so
it is a simple matter of updating the `NewSSODiagContext()` call sites
to pass the `auth.Server` instead of use it as a received. A couple of
test cases that wanted to stub out the backend now do so via the adapter
func `SSODiagServiceFunc`, implemented in the same vein as
`http.Handler/http.HandlerFunc`.

This will be needed to move some SSO plugins out of the OSS repo.