auth: Export unstable field of auth.Server
Export the `unstable` field of `auth.Server` so it can be accessed by
the SAML connector code when it is moved out-of-tree.

Further refactoring is possible to remove the unstable service
altogether, since it contains only one service -
`AssertionReplayService` and that is used only by the SAML connector.
However there is also an API endpoint that calls the service. That is
slated to be removed in v12, so the further refactoring can wait until
that is addressed.